Initial typescript scaffold
