+++ package.json
@@ -31,9 +31,14 @@
     "svelte": "^3.50.1"
   },
   "devDependencies": {
+    "@rollup/plugin-typescript": "^9.0.2",
     "@sveltejs/vite-plugin-svelte": "^1.0.9",
+    "@tsconfig/svelte": "^3.0.0",
+    "svelte-check": "^2.9.2",
     "svelte-color-picker": "^1.0.7",
     "svelte-colorpick": "^0.1.7",
+    "svelte-preprocess": "^4.10.7",
+    "typescript": "^4.8.4",
     "vite": "^3.1.8",
     "vite-plugin-banner": "^0.5.0"
   }
