+++ src/components/Chart.svelte
@@ -9,7 +9,7 @@ import DataHub from '../core/dataHub.js'
 import MetaHub from '../core/metaHub.js'
 import Scan from '../core/dataScanner.js'
 import Events from '../core/events.js'
-import Const from '../stuff/constants.js'
+import Const from '../stuff/constants'
 import Utils from '../stuff/utils.js'
 import Layout from '../core/layout.js'
 import Context from '../stuff/context.js'
